sela/requirements.txt
anthropic>=0.50.0
yfinance>=0.2.40
flask>=2.3
python-telegram-bot[job-queue]>=20.7
